+++ UI/src/com/kotcrab/vis/ui/widget/LinkLabel.java
@@ -19,8 +19,10 @@ package com.kotcrab.vis.ui.widget;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input.Buttons;
 import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.graphics.Cursor.SystemCursor;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
+import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
@@ -97,6 +99,18 @@ public class LinkLabel extends VisLabel {
 				} else {
 					listener.clicked(url.toString());
 				}
+			}
+
+			@Override
+			public void enter (InputEvent event, float x, float y, int pointer, Actor fromActor) {
+				super.enter(event, x, y, pointer, fromActor);
+				Gdx.graphics.setSystemCursor(SystemCursor.Hand);
+			}
+
+			@Override
+			public void exit (InputEvent event, float x, float y, int pointer, Actor toActor) {
+				super.exit(event, x, y, pointer, toActor);
+				Gdx.graphics.setSystemCursor(SystemCursor.Arrow);
 			}
 		});
 	}